services/vibe/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
httpx==0.27.0
aiosqlite==0.20.0

# Data Sources
pykrx>=1.2.4
yfinance>=0.2.66
finance-datareader>=0.9.90

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0

# XML Security
defusedxml>=0.7.1

# Scheduling
APScheduler>=3.10.0,<4.0
pytz>=2024.1

# File Upload
python-multipart>=0.0.9

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# LLM Providers (optional, required when LLM features enabled)
anthropic>=0.40.0
openai>=1.40.0
